backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
asgiref==3.7.2
attrs==23.1.0
black==23.12.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.9.0
coverage==7.8.0
Django==5.0
django-cors-headers==4.3.1
django-countries==7.6.1
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
exceptiongroup==1.2.0
Faker==21.0.0
frozenlist==1.5.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
propcache==0.3.0
psycopg2-binary==2.9.9
pur==7.3.1
pycodestyle==2.11.1
pydotplus==2.0.2
pyflakes==3.1.0
PyJWT==2.8.0
pyparsing==3.2.3
pytest==7.4.3
pytest-cov==6.0.0
pytest-django==4.7.0
pytest-lazy-fixture==0.6.3
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-stdnum==1.20
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
ruff==0.11.2
setuptools==76.0.0
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
wheel==0.45.1
yarl==1.18.3
